Modify generic repository interfaces
Modification of generic repository interfaces in all classes of the SERVICE package.

The generic repository interface is updated to a unique one for each type of class. This change is implemented to provide unique and personalized methods for each repository, tailored to the specific requirements of the classes in the DOMAIN package.